optimised code and added types
