+++ cmd/armv/app/app.go
@@ -29,24 +29,26 @@ import (
 	"time"
 
 	"github.com/AaronSaikovski/armv/cmd/armv/poller"
-	"github.com/AaronSaikovski/armv/cmd/armv/types"
 	"github.com/AaronSaikovski/armv/internal/pkg/auth"
 	"github.com/AaronSaikovski/armv/internal/pkg/resourcegroups"
 	"github.com/AaronSaikovski/armv/internal/pkg/resources"
 	"github.com/AaronSaikovski/armv/internal/pkg/validation"
 	"github.com/AaronSaikovski/armv/pkg/utils"
+	"github.com/Azure/azure-sdk-for-go/sdk/azidentity"
 	"github.com/logrusorgru/aurora"
 )
 
 var (
-	args              utils.Args
-	azureResourceInfo = types.AzureResourceInfo{}
+	args utils.Args
 )
 
 // run - main run method
 func Run(versionString string) error {
 
 	startTime := time.Now()
+	defer func() {
+		fmt.Printf("Elapsed time: %.2f seconds\n", time.Since(startTime).Seconds())
+	}()
 
 	//set the version build info
 	args.SetVersion(versionString)
@@ -59,9 +61,11 @@ func Run(versionString string) error {
 	/* ********************************************************************** */
 
 	//populate the AzureResourceInfo struct
-	azureResourceInfo.SourceSubscriptionId = args.SourceSubscriptionId
-	azureResourceInfo.SourceResourceGroup = args.SourceResourceGroup
-	azureResourceInfo.TargetResourceGroup = args.TargetResourceGroup
+	azureResourceMoveInfo := validation.AzureResourceMoveInfo{
+		SourceSubscriptionId: args.SourceSubscriptionId,
+		SourceResourceGroup:  args.SourceResourceGroup,
+		TargetResourceGroup:  args.TargetResourceGroup,
+	}
 
 	/* ********************************************************************** */
 
@@ -72,30 +76,39 @@ func Run(versionString string) error {
 	/* ********************************************************************** */
 
 	// Get default cred
-	cred, err := auth.GetAzureDefaultCredential()
+	cred, err := func() (*azidentity.DefaultAzureCredential, error) {
+		return azidentity.NewDefaultAzureCredential(nil)
+	}()
+
 	if err != nil {
-		return err
+		return fmt.Errorf("failed to get Azure default credential: %w", err)
 	}
 
 	/* ********************************************************************** */
 	// check we are logged into the Azure source subscription
-	if !auth.GetLogin(ctx, args.SourceSubscriptionId) {
-		return fmt.Errorf("you are not logged into the azure subscription '%s', please login and retry operation", azureResourceInfo.SourceSubscriptionId)
+	login, err := auth.CheckLogin(ctx, cred, args.SourceSubscriptionId)
+
+	if err != nil {
+		return fmt.Errorf("getlogin error: %w", err)
 	}
-	fmt.Println(aurora.Sprintf(aurora.Yellow("Logged into Subscription Id: %s\n"), azureResourceInfo.SourceSubscriptionId))
+
+	if !login {
+		return fmt.Errorf("you are not logged into the azure subscription '%s', please login and retry operation", azureResourceMoveInfo.SourceSubscriptionId)
+	}
+	fmt.Println(aurora.Sprintf(aurora.Yellow("Logged into Subscription Id: %s\n"), azureResourceMoveInfo.SourceSubscriptionId))
 
 	/* ********************************************************************** */
 
 	//Get the resource group client
-	resourceGroupClient, err := resourcegroups.GetResourceGroupClient(cred, azureResourceInfo.SourceSubscriptionId)
+	resourceGroupClient, err := resourcegroups.GetResourceGroupClient(cred, azureResourceMoveInfo.SourceSubscriptionId)
 	if err != nil {
-		return err
+		return fmt.Errorf("failed to get resource group client: %w", err)
 	}
 
 	/* ********************************************************************** */
 
 	// check source and destination resource groups exists
-	srcRsgExists, err := resourcegroups.CheckResourceGroupExists(ctx, resourceGroupClient, azureResourceInfo.SourceResourceGroup)
+	srcRsgExists, err := resourcegroups.CheckResourceGroupExists(ctx, resourceGroupClient, azureResourceMoveInfo.SourceResourceGroup)
 	if err != nil {
 		return err
 	}
@@ -106,7 +119,7 @@ func Run(versionString string) error {
 	/* ********************************************************************** */
 
 	// check destination and destination resource groups exists
-	dstRsgExists, err := resourcegroups.CheckResourceGroupExists(ctx, resourceGroupClient, azureResourceInfo.TargetResourceGroup)
+	dstRsgExists, err := resourcegroups.CheckResourceGroupExists(ctx, resourceGroupClient, azureResourceMoveInfo.TargetResourceGroup)
 	if err != nil {
 		return err
 	}
@@ -117,7 +130,7 @@ func Run(versionString string) error {
 	/* ********************************************************************** */
 
 	// Get resource client
-	resourcesClient, err := resources.GetResourcesClient(cred, azureResourceInfo.SourceSubscriptionId)
+	resourcesClient, err := resources.GetResourcesClient(cred, azureResourceMoveInfo.SourceSubscriptionId)
 	if err != nil {
 		return err
 	}
@@ -125,42 +138,35 @@ func Run(versionString string) error {
 	/* ********************************************************************** */
 
 	// Get all resource IDs from source resource group
-	azureResourceInfo.ResourceIds, err = resources.GetResourceIds(ctx, resourcesClient, azureResourceInfo.SourceResourceGroup)
+	azureResourceMoveInfo.ResourceIds, err = resources.GetResourceIds(ctx, resourcesClient, azureResourceMoveInfo.SourceResourceGroup)
 	if err != nil {
-		return err
+		return fmt.Errorf("failed to get resource IDs: %w", err)
 	}
 	/* ********************************************************************** */
 
 	// get the target resource group ID
-	targetResourceGroupId, err := resourcegroups.GetResourceGroupId(ctx, resourceGroupClient, azureResourceInfo.TargetResourceGroup)
+	azureResourceMoveInfo.TargetResourceGroupId, err = resourcegroups.GetResourceGroupId(ctx, resourceGroupClient, azureResourceMoveInfo.TargetResourceGroup)
 	if err != nil {
-		return err
+		return fmt.Errorf("failed to get target resource group ID: %w", err)
 	}
 
 	/* ********************************************************************** */
 
 	//Validate resources - return runtime poller
-	resp, err := validation.ValidateMove(ctx, args.SourceSubscriptionId, args.SourceResourceGroup, azureResourceInfo.ResourceIds, targetResourceGroupId)
+	resp, err := azureResourceMoveInfo.ValidateMove(ctx, cred)
 	if err != nil {
-		return err
+		return fmt.Errorf("failed to validate resource move: %w", err)
 	}
 
 	/* ********************************************************************** */
 
 	// Poll the API and show a status...this is a blocking call
-	pollResp, err := poller.PollApi(ctx, resp)
-	if err != nil {
-		return err
-	}
-
-	//Show response output
-	respErr := poller.PollResponse(pollResp)
-	if respErr != nil {
-		return respErr
+	pollErr := poller.PollApi(ctx, resp)
+	if pollErr != nil {
+		return fmt.Errorf("failed to poll API: %w", pollErr)
 	}
 
 	/* ********************************************************************** */
 
-	fmt.Printf("Elapsed time: %.2f seconds\n", time.Since(startTime).Seconds())
 	return nil
 }
