+++ cmd/armv/version.txt
@@ -1,1 +1,1 @@
-v0.2.0
+v0.3.0
